# [ hdot ] Use precision-enhanced hdot