**hdot : vec(1 x K ) x vec(K X 1 ) -> 1 (scalar)**

- Previous hdot was using full-fp16.
- Since this is also one of dimension-shrinking computations, should use inter-fp32 values to enhance precision.
- This has not been detected due to small dimension Tensor usage in unittest. Add bigger-sized test case accordingly.

### A. Accuracy w.r.t cblas-f32 (mse)
| dim | hdot-f16 (previous) | hdot-f16f32 (now) |
| --- | --- | --- |
| 1024 | 0.250061 | 3.72529e-09 |
| 2048 | 2.07308 | 0.00362171 |
| 4096 | 11.8475 | 0.195379 |

### B. Latency
> Since this is nanosec-unit result, almost trivial

| dim | hdot-f16 | hdot-f16f32 | cblas-f32 |
| --- | --- | --- | --- |
| 1024 | 261 ns | 280 ns | 347 ns |
| 2048 | 328 ns | 359 ns | 1473 ns |
| 4096 | 508 ns | 562 ns | 909 ns |

Conclusion : Better accuracy with almost no latency deterioration

**Self evaluation:**
1. Build test:     [X]Passed [ ]Failed [ ]Skipped
2. Run test:     [X]Passed [ ]Failed [ ]Skipped